Adds ability to reuse an app so that tests are less expensive

+++ src/test/groovy/tests/support/FastlyToHeroku.groovy
@@ -11,8 +11,10 @@ class FastlyToHeroku {
 
         if (!service) {
             service = fastly.createService(appName, "${appName}.herokuapp.com")
-            Runtime.runtime.addShutdownHook {
-                fastly.destroyService(service.id)
+            if (!System.getenv('APP_NAME')) {
+                Runtime.runtime.addShutdownHook {
+                    fastly.destroyService(service.id)
+                }
             }
         }
